requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.1.4
django-allauth==65.3.0
django-ninja==1.3.0
django-redis==5.4.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-tinymce==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.10.1
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
django-mptt~=0.16.0
drf-yasg~=1.21.8
django-admin-sortable2~=2.2.4
django-filter~=24.3